refactor: Update UpdateInstitutionUsecase to handle social media links and countries

+++ src/modules/update_institution/app/update_institution_usecase.ts
@@ -72,6 +72,23 @@ export class UpdateInstitutionUsecase {
       }
     }
 
+    if (body.social_medias) {
+      body.social_medias.forEach((social_media: { id: number, link: string }) => {
+        if (!social_media.id) {
+          throw new MissingParameter("Social Media Id");
+        }
+        if (!social_media.link) {
+          throw new MissingParameter("Social Media Link");
+        }
+      });
+    }
+
+    if (body.countries) {
+      body.countries.forEach((country_id: number) => {
+        return { id: country_id };
+      });
+    }
+
     institution.name = body.name || institution.name;
     institution.description = body.description || institution.description;
     institution.countries = body.country || institution.countries;
@@ -79,15 +96,17 @@ export class UpdateInstitutionUsecase {
     institution.images = body.images || institution.images;
     institution.social_medias = body.social_medias || institution.social_medias;
 
-    if (body.images) {
-      await this.bucket.delete_folder(`institution/${institution.id}/`).then(() => {
-        institution.images.map(async (image: string, index: number) => {
-          const content_type = image.split(';')[0].split(':')[1];
-          const image_key = `institution/${institution.id}/${index}.${content_type.split('/')[1]}`;
-          const image_buffer = Buffer.from(image.split(',')[1], 'base64');
-          institution.images[index] = await this.bucket.upload_image(image_key, image_buffer, content_type);
+    if (process.env.STAGE !== 'test') {
+      if (body.images) {
+        await this.bucket.delete_folder(`institution/${institution.id}/`).then(() => {
+          institution.images.map(async (image: string, index: number) => {
+            const content_type = image.split(';')[0].split(':')[1];
+            const image_key = `institution/${institution.id}/${index}.${content_type.split('/')[1]}`;
+            const image_buffer = Buffer.from(image.split(',')[1], 'base64');
+            institution.images[index] = await this.bucket.upload_image(image_key, image_buffer, content_type);
+          });
         });
-      });
+      }
     }
 
     await this.institution_repo.update_institution(institution);
